docs(#1491): state the refusal as ONE derived question, and split the unit file
The module doc described five node-local causes as a checklist; there is
now one derived cause (the decoder does not give this node's members back)
plus EMPTY-CONTAINER, which is not a recovery question, plus the
whole-cell UNBALANCED bracket. Every site that described the old causes,
the 'structural character' whole-cell wording or the retired
'ambiguity 0/1/2' numbering is updated with it.

Also records what the derived question is NOT: refusing on
'could another value have rendered these bytes' would refuse every
multi-member text collection, because unquoted members make ['a','b'] and
['a, b'] the same bytes. That inherent limit is accepted and stated.

The unit cases move to golden_csv_container_tests.rs (campsite rule,
#1135): 1629 -> 1046 + 593.
